+++ src/Frontend/Dfe.Complete/Pages/Projects/InternalContacts/EditAddedByUser.cshtml.cs
@@ -21,9 +21,9 @@ public class EditAddedByUser(ISender sender, ErrorService errorService, ILogger<
 
     public UserDto AddedByUser { get; set; } = default!;
     
-    public override async Task<IActionResult> OnGet()
+    public override async Task<IActionResult> OnGetAsync()
     {
-        await base.OnGet();
+        await base.OnGetAsync();
         var addedByUserQuery = new GetUserByIdQuery(Project.RegionalDeliveryOfficerId);
         var addedbyResult = await _sender.Send(addedByUserQuery);
         if (addedbyResult is { IsSuccess: true, Value: not null })
@@ -45,7 +45,7 @@ public class EditAddedByUser(ISender sender, ErrorService errorService, ILogger<
         if (!ModelState.IsValid)
         {
             errorService.AddErrors(ModelState);
-            return await OnGet();
+            return await OnGetAsync();
         }
         
         var addedByUserQuery = new GetUserByEmailQuery(Email);
@@ -61,6 +61,6 @@ public class EditAddedByUser(ISender sender, ErrorService errorService, ILogger<
 
         logger.LogError("Email not found or not assignable - {Email}", addedByUserQuery.Email);
         ModelState.AddModelError("Email", "Email is not assignable");
-        return await OnGet();
+        return await OnGetAsync();
     }
 }